+++ nimble/host/src/ble_sm_priv.h
@@ -437,7 +437,7 @@ void ble_sm_deinit(void);
 #endif
 
 struct ble_l2cap_chan *ble_sm_create_chan(uint16_t handle);
-void *ble_sm_cmd_get(uint8_t opcode, size_t len, struct os_mbuf **txom);
+void *ble_sm_cmd_get(uint8_t opcode, uint16_t len, struct os_mbuf **txom);
 int ble_sm_tx(uint16_t conn_handle, struct os_mbuf *txom);
 
 #if MYNEWT_VAL(BLE_STATIC_TO_DYNAMIC)
